Add author field for Message (OneToMany mapping)

+++ src/main/java/ru/kromarong/twit/repos/MessageRepo.java
@@ -1,11 +1,12 @@
 package ru.kromarong.twit.repos;
 
-import org.springframework.data.repository.CrudRepository;
 import ru.kromarong.twit.domain.Message;
+import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
 
     List<Message> findByTag(String tag);
+
 }
